REFACT-FEAT: Introdução do domínio Workspace e persistência de enums como strings
- Adição das entidades Workspace e MembroWorkspace, do enum TipoWorkspace e das configurações do EF Core
- Criação de migrações para as tabelas Workspaces e MembrosWorkspace, incluindo restrições (constraints)
- Persistência de enums como strings descritivas utilizando EnumStringConverter
- Adição de migrações para converter colunas de enum de inteiro (int) para string, com SQL para migração de dados
- Refatoração da base de código para utilizar enums fortemente tipados nas entidades e consultas
- Atualização da documentação referente ao modelo de workspace e às alterações na persistência de enums
- Adição de métodos em EnumExtensions para conversão rigorosa de descrição para enum

+++ AtronPlatform/Modules/Tracker/Application/Mapping/SolicitacaoObtencaoTarefaMapping.cs
@@ -19,7 +19,7 @@ namespace Application.Mapping
             {
                 Id = solicitacao.Id,
                 TarefaId = solicitacao.TarefaId,
-                Status = solicitacao.Status,
+                Status = (int)solicitacao.Status,
                 DataSolicitacao = solicitacao.DataSolicitacao,
                 DataDecisao = solicitacao.DataDecisao,
                 Tarefa = solicitacao.Tarefa.MapToDto(_tarefaMapper),
